refactor(cli): move get into provider-generator
