+++ packages/@cdktf/cli-core/src/lib/index.ts
@@ -18,7 +18,6 @@ export {
 } from "./error-reporting";
 export { CdktfStack } from "./cdktf-stack";
 export { init, Project, InitArgs } from "./init";
-export { get, GetStatus } from "./get";
 export { SynthesizedStack } from "./synth-stack";
 export { CdktfProject, ProjectUpdate } from "./cdktf-project";
 export { watch, State as WatchState } from "./watch";
